+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Version 0.2.0
 --------------
 
++ Add cache for ``url_for``
 + Add cache for ``static_url``
 + Don't generate analytics codes when running ``catsup server``